Refactored out of openbazaard.py into daemon.py
